felisha-agent/requirements.txt
livekit-agents>=0.8,<0.9
livekit-plugins-deepgram>=0.6,<0.7
httpx>=0.27,<0.28
python-dotenv>=1.0,<2.0
